better examples and naming

+++ index.js
@@ -12,7 +12,7 @@ var spawn = require('child_process').spawn,
     events = require('events'),
     util = require('util');
 
-module.exports = function Sound(filename) {
+module.exports = function Media(filename) {
     events.EventEmitter.call(this);
     this.filename = filename;
 };
@@ -22,6 +22,7 @@ util.inherits(module.exports, events.EventEmitter);
 module.exports.prototype.play = function (options) {
     this.stopped = false;
     var args = [this.filename];
+
     for(var prop in options) {
         if(options.hasOwnProperty(prop)){
             args.unshift('-'+prop, options[prop] );
@@ -34,18 +35,25 @@ module.exports.prototype.play = function (options) {
         }
     }.bind(this));
 };
+
 module.exports.prototype.stop = function () {
     this.stopped = true;
-    this.process.kill('SIGTERM');
+    if(this.process){
+        this.process.kill('SIGTERM');
+    }
     this.emit('stop');
 };
 module.exports.prototype.pause = function () {
     if (this.stopped) return;
-    this.process.kill('SIGSTOP');
+    if(this.process){
+        this.process.kill('SIGSTOP');
+    }
     this.emit('pause');
 };
 module.exports.prototype.resume = function () {
     if (this.stopped) return this.play();
-    this.process.kill('SIGCONT');
+    if(this.process){
+        this.process.kill('SIGCONT');
+    }
     this.emit('resume');
 };
